backend/requirements.txt
# Irerero Backend — Python Dependencies
# Install: pip install -r requirements.txt

# Django core
django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
django-cors-headers==4.4.0

# Database
psycopg2-binary==2.9.10

# Password hashing — NFR-020
argon2-cffi==23.1.0
bcrypt==4.2.0

# Async / WebSocket — architecture §5.4
# Channels 4 does not patch `manage.py runserver`; use Daphne so /ws/** works.
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2

# Background tasks — Celery
celery==5.4.0
redis==5.1.1
django-celery-beat==2.7.0

# AI / ML — architecture §7
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.3
shap==0.46.0
python-dateutil==2.9.0

# PDF generation — WeasyPrint (referral slips, reports)
weasyprint==62.3

# Image handling — child photos
Pillow==10.4.0

# SMS — Africa's Talking (production)
africastalking==1.2.8

# FCM push notifications
firebase-admin==6.5.0

# Excel/CSV export — FR-073
openpyxl==3.1.5
